Pass int to is* character classification functions as per SUSv3
This silences warnings during compilation on Solaris and HP-UX
(Mantis #1177).

+++ src/conversion.c
@@ -1322,7 +1322,7 @@ static int IsSpace(char *remainder)
 
     for (sp = remainder; *sp != '\0'; sp++)
     {
-        if (!isspace(*sp))
+        if (!isspace((int)*sp))
         {
             return false;
         }
